[word-detective] Use cxx 17
This commit sets the build to use cxx 17
standard. Additionally, it fixes some
potential causes of undefined behaviour in
the code.

+++ app/word-detective/include/word-detective-api.h
@@ -57,9 +57,9 @@ std::string path_from_word(const std::string& word);
 
 void get_all_puzzles(std::list<Puzzle::Puzzle>& list_of_puzzles,
                      const Datastr::Brick& brick, int num_letters);
-json all_puzzles(const Datastr::Brick& brick, int num_letters, int min_words);
-json random_puzzle(const Datastr::Brick& brick, int num_letters, int min_words);
-json puzzle_by_index(const Datastr::Brick& brick, int num_letters,
-                     int min_words, int index);
+json all_puzzles(const Datastr::Brick& brick, size_t num_letters, size_t min_words);
+json random_puzzle(const Datastr::Brick& brick, size_t num_letters, size_t min_words);
+json puzzle_by_index(const Datastr::Brick& brick, size_t num_letters,
+                     size_t min_words, int index);
 
 #endif
